+++ src/index.ts
@@ -1,6 +1,8 @@
 /** Support for package.json `main` entrypoint for backwards compatibility. */
 
 export * from './core.js';
+
+export * from './config/index.js';
 export * from './fs/fs.client.js';
 export * from './http/http.client.js';
 export * from './http/http.server.js';